move tests to better spot
