+++ nilearn/tests/test_package_installation.py
@@ -1,17 +1,6 @@
-import pytest
 from nilearn.version import (
     REQUIRED_MODULE_METADATA, _compare_version
 )
-
-
-@pytest.mark.parametrize("version_a,operator,version_b",
-                         [('0.1.0', '>', '0.0.1'),
-                          ('0.1.0', '>=', '0.0.1'),
-                          ('0.1', '==', '0.1.0'),
-                          ('0.0.0', '<', '0.1.0'),
-                          ('1.0', '!=', '0.1.0')])
-def test_compare_version(version_a, operator, version_b):
-    assert _compare_version(version_a, operator, version_b)
 
 
 def test_required_package_installation():
